♻️ #103 refactor: auth관련 타입정리 안쓰는 api 삭제
