--- lib/api/getTodos.ts
@@ -1,26 +1,0 @@
-import { GetTodosResponse } from '../types/todos';
-
-export const getTodos = async ({
-  goalId,
-  done,
-  cursor,
-  size = 20,
-}: {
-  goalId?: number;
-  done?: boolean;
-  cursor?: number;
-  size?: number;
-}): Promise<GetTodosResponse> => {
-  const params = new URLSearchParams();
-  if (goalId) params.append('goalId', goalId.toString());
-  if (done !== undefined) params.append('done', done.toString());
-  if (cursor) params.append('cursor', cursor.toString());
-  if (size) params.append('size', size.toString());
-
-  const response = await fetch(`/4-4-dev/todos?${params.toString()}`, {
-    method: 'GET',
-    headers: { 'Content-Type': 'application/json' },
-  });
-  if (!response.ok) throw new Error('Failed to fetch todos');
-  return response.json();
-};
