+++ lib/api/signUp.ts
@@ -1,20 +1,5 @@
 import { API_BASE_URL } from '@/constants/api';
-
-interface SignupFormData {
-  name: string;
-  email: string;
-  password: string;
-}
-
-interface SignupResponse {
-  user: {
-    id: number;
-    name: string;
-    email: string;
-    createdAt: string;
-    updatedAt: string;
-  };
-}
+import { SignupFormData, SignupResponse } from '../types/auth';
 
 export const signUp = async (data: SignupFormData): Promise<SignupResponse> => {
   const response = await fetch(`${API_BASE_URL}/user`, {
